abstracted the function for getting latest club member list to a lib function
