+++ lib/ethereum.ts
@@ -3,6 +3,11 @@ export const signInMessage =
   "Welcome to frens!\n\nYou are one step away from investing cryptocurrencies with your friends.\n\nClick to sign in and accept the frens Terms of Service\n\nThis request will not trigger a blockchain transaction or cost any gas fees.";
 import axios from "axios";
 import { BigNumber, ethers } from "ethers";
+import { doc, updateDoc } from "firebase/firestore";
+import _ from "lodash";
+import { adminFirestore } from "../firebase/firebaseAdmin";
+import { clientFireStore } from "../firebase/firebaseClient";
+import { IClubInfo } from "../pages/clubs/[id]";
 import { getChainData } from "./chains";
 
 interface IHolderBalanceInfo {
@@ -22,6 +27,16 @@ interface ITransferEvent {
   value: string;
   transaction_index: number;
   log_index: number;
+}
+
+export async function getLatestBlockNumber() {
+  // STEP 2: Get the transfer events ellapsed from last time the club is retrieved
+  const rpcUrl = getChainData(
+    parseInt(process.env.NEXT_PUBLIC_ACTIVE_CHAIN_ID!)
+  ).rpc_url;
+  const provider = new ethers.providers.JsonRpcProvider(rpcUrl);
+  const currentBlock = await provider.getBlockNumber();
+  return currentBlock;
 }
 
 // function to verfiy signature with user address
@@ -173,4 +188,66 @@ export async function getUsdPrice(tokenAddress?: string): Promise<number> {
     console.log("Fetch usd price error: ", err);
   }
   return usdPrice;
+}
+
+export async function getClubMemberAddress (clubInfo: IClubInfo, id: string) {
+  // STEP 1: Fetch the last updated club member list
+  let _club_members: { [k: string]: number };
+  if (clubInfo.club_members) {
+    _club_members = clubInfo.club_members;
+  } else {
+    _club_members = {
+      "0x0000000000000000000000000000000000000000": 0,
+    };
+  }
+
+  // STEP 2: Get the transfer events ellapsed from last time the club is retrieved
+  const rpcUrl = getChainData(
+    parseInt(process.env.NEXT_PUBLIC_ACTIVE_CHAIN_ID!)
+  ).rpc_url;
+  const provider = new ethers.providers.JsonRpcProvider(rpcUrl);
+  const currentBlock = await provider
+    .getBlockNumber();
+  // console.log(`Querying transactions from ${clubInfo.last_retrieved_block} to ${currentBlock}`);
+  const MORALIS_API_KEY = process.env.NEXT_PUBLIC_MORALIS_KEY;
+  const options = {
+    method: "GET",
+    url: "https://deep-index.moralis.io/api/v2/erc20/%address%/transfers".replace(
+      "%address%",
+      clubInfo.club_token_address!
+    ),
+    params: {
+      chain: getChainData(parseInt(process.env.NEXT_PUBLIC_ACTIVE_CHAIN_ID!))
+        .network,
+      from_block:
+        clubInfo.last_retrieved_block && clubInfo.last_retrieved_block,
+      to_block: currentBlock,
+    },
+    headers: { accept: "application/json", "X-API-Key": MORALIS_API_KEY },
+  };
+
+  // STEP 3: Update the club member list based on new transfer events
+  const transferEvents = await axios
+    .request(options)
+    .then((response) => response.data)
+    .then((data) => data.result);
+  // console.log(transferEvents);
+  transferEvents.forEach((event: ITransferEvent) => {
+    // Create a new member object if the existing club member object does not have the addresses
+    if (!(event.from_address in _club_members)) {
+      _club_members[event.from_address] = 0;
+    }
+    if (!(event.to_address in _club_members)) {
+      _club_members[event.to_address] = 0;
+    }
+    // Update the balance of each club member
+    _club_members[event.from_address] -= parseInt(event.value);
+    _club_members[event.to_address] += parseInt(event.value);
+  });
+  // Purge all addresses with <=0 balance
+  _club_members = _.pickBy(_club_members, function (value) {
+    return value > 0;
+  });
+  // console.log('New club member list: ', _club_members)
+  return _club_members
 }
