+++ pages/clubs/[id]/index.tsx
@@ -1,10 +1,7 @@
 import React, { lazy, ReactElement } from "react";
 import AppLayout from "../../../layout/AppLayout";
 import { NextPageWithLayout } from "../../_app";
-import {
-  adminAuth,
-  adminFirestore,
-} from "../../../firebase/firebaseAdmin";
+import { adminAuth, adminFirestore } from "../../../firebase/firebaseAdmin";
 import { InferGetServerSidePropsType } from "next";
 import ClubDetails from "../../../components/ClubDetails/ClubDetails";
 import ClubMembers from "../../../components/ClubMembers/ClubMembers";
@@ -17,7 +14,8 @@ import _ from "lodash";
 import nookies from "nookies";
 import NotAuthed from "../../../components/NotAuthed/NotAuthed";
 import {
-  getUsdPrice,
+  getClubMemberAddress,
+  getLatestBlockNumber,
   verifyClubHolding,
 } from "../../../lib/ethereum";
 import NotVerified from "../../../components/NotVerified/NotVerified";
@@ -160,73 +158,20 @@ export const getServerSideProps = async (context: any) => {
     }
   };
 
-  const fetchMemberInfo = async (clubInfo: IClubInfo): Promise<IMemberInfoData[]> => {
-    // STEP 1: Fetch the last updated club member list
-    let _club_members: { [k: string]: number };
-    if (clubInfo.club_members) {
-      _club_members = clubInfo.club_members;
-    } else {
-      _club_members = {
-        "0x0000000000000000000000000000000000000000": 0,
-      };
-    }
-
-    // STEP 2: Get the transfer events ellapsed from last time the club is retrieved
-    const rpcUrl = getChainData(
-      parseInt(process.env.NEXT_PUBLIC_ACTIVE_CHAIN_ID!)
-    ).rpc_url;
-    const provider = new ethers.providers.JsonRpcProvider(rpcUrl);
-    const currentBlock = await provider
-      .getBlockNumber();
-    // console.log(`Querying transactions from ${clubInfo.last_retrieved_block} to ${currentBlock}`);
-    const MORALIS_API_KEY = process.env.NEXT_PUBLIC_MORALIS_KEY;
-    const options = {
-      method: "GET",
-      url: "https://deep-index.moralis.io/api/v2/erc20/%address%/transfers".replace(
-        "%address%",
-        clubInfo.club_token_address!
-      ),
-      params: {
-        chain: getChainData(parseInt(process.env.NEXT_PUBLIC_ACTIVE_CHAIN_ID!))
-          .network,
-        from_block:
-          clubInfo.last_retrieved_block && clubInfo.last_retrieved_block,
-        to_block: currentBlock,
-      },
-      headers: { accept: "application/json", "X-API-Key": MORALIS_API_KEY },
-    };
-
-    // STEP 3: Update the club member list based on new transfer events
-    const transferEvents = await axios
-      .request(options)
-      .then((response) => response.data)
-      .then((data) => data.result);
-    // console.log(transferEvents);
-    transferEvents.forEach((event: ITransferEvent) => {
-      // Create a new member object if the existing club member object does not have the addresses
-      if (!(event.from_address in _club_members)) {
-        _club_members[event.from_address] = 0;
-      }
-      if (!(event.to_address in _club_members)) {
-        _club_members[event.to_address] = 0;
-      }
-      // Update the balance of each club member
-      _club_members[event.from_address] -= parseInt(event.value);
-      _club_members[event.to_address] += parseInt(event.value);
-    });
-    // Purge all addresses with <=0 balance
-    _club_members = _.pickBy(_club_members, function (value) {
-      return value > 0;
-    });
-    // console.log('New club member list: ', _club_members)
-
-    // STEP 4: Replace existing club members with updated club members list
-    const result = await adminFirestore.collection("clubs").doc(id).update({
+  const fetchMemberInfo = async (
+    clubInfo: IClubInfo
+  ): Promise<IMemberInfoData[]> => {
+    // STEP 1: Fetch the latest club member list
+    const _club_members = await getClubMemberAddress(clubInfo, id);
+    
+    // STEP 2: Update the club member list
+    const currentBlock = await getLatestBlockNumber()
+    const result = adminFirestore.collection("clubs").doc(id).update({
       club_members: _club_members,
       last_retrieved_block: currentBlock,
     });
 
-    // STEP 5: Fetch club members info
+    // STEP 3: Fetch club members info by the updated club member list
     let memberInfo = [] as IMemberInfoData[];
     await Promise.all(
       Object.keys(_club_members).map(async (uid) => {
@@ -239,7 +184,7 @@ export const getServerSideProps = async (context: any) => {
         });
       })
     );
-    return memberInfo
+    return memberInfo;
   };
 
   if (!cookies.token) {
@@ -274,11 +219,11 @@ export const getServerSideProps = async (context: any) => {
       );
 
       // Step 4: fetch club members
-      let memberInfo = [] as IMemberInfoData[]
+      let memberInfo = [] as IMemberInfoData[];
       try {
         memberInfo = await fetchMemberInfo(clubInfo);
       } catch (err) {
-        console.log(err)
+        console.log(err);
       }
 
       return {
@@ -317,10 +262,13 @@ const Dashboard: NextPageWithLayout<any> = ({
             {/* TODO: have a global state setting for whether to show club or me balance */}
             <ClubBalance />
             {/* Portfolio */}
-            <Portfolio data={serverProps.balance!} clubWalletAddress={serverProps.clubInfo?.club_wallet_address!}/>
+            <Portfolio
+              data={serverProps.balance!}
+              clubWalletAddress={serverProps.clubInfo?.club_wallet_address!}
+            />
           </div>
           {/* Right panel */}
-          <WidgetSection data={serverProps.clubInfo!}/>
+          <WidgetSection data={serverProps.clubInfo!} />
           {/* FOR TESTING SPLITTING */}
           <Splitting data={serverProps.clubInfo!} />
         </div>
